Attempt for Edit task TODO:FIX null Id

+++ src/saveTask.js
@@ -1,5 +1,8 @@
 export default function saveTask(task){
 
+    const confirmAddTask = document.getElementById('confirmButton-task');
+
+
     let taskId = 0;
 
     if (!localStorage.getItem('taskId')) {
@@ -9,10 +12,17 @@ export default function saveTask(task){
         taskId = parseInt(localStorage.getItem('taskId'));
     }
 
-    task.id = taskId.toString();
+    console.log(task.id);
+    console.log(taskId);
 
-    localStorage.setItem(taskId.toString(), JSON.stringify(task));
-
+    if(confirmAddTask.dataset.type = 'edit'){
+        
+        localStorage.setItem(task.id, JSON.stringify(task));
+    }
+    if(confirmAddTask.dataset.type = 'add'){
+        task.id = taskId.toString();
+        localStorage.setItem(taskId.toString(), JSON.stringify(task));
+    }
     taskId++;
     localStorage.setItem('taskId', taskId);
 }
